+++ internal/model/withdrawals.go
@@ -4,6 +4,6 @@ import "time"
 
 type GetWithdrawalsResponse struct {
 	Order       string    `json:"order"`
-	Sum         int       `json:"sum"`
+	Sum         float32   `json:"sum"`
 	ProcessedAt time.Time `json:"processed_at"`
 }
